continued class refactoring
