+++ src/pysiral/scripts/cfgsetdir.py
