+++ src/pysiral/scripts/cfgupdate.py
