+++ src/pysiral/scripts/l3proc.py
@@ -13,7 +13,7 @@ from dateperiods import DatePeriod
 from loguru import logger
 
 from pysiral import psrlcfg
-from pysiral.core.config import DefaultCommandLineArguments
+from pysiral.scripts.parser_items import DefaultCommandLineArguments
 from pysiral.core.datahandler import L2iDataHandler
 from pysiral.core.legacy_classes import DefaultLoggingClass, ErrorStatus
 from pysiral.l3proc import (Level3GridDefinition, Level3OutputHandler,
